+++ src/components/LogCard/MainPageLogCardList.tsx
@@ -2,7 +2,7 @@ import MainPagination from '@/components/Pagination/MainPagination';
 import useLogList from '@/hooks/queries/log/useLogList';
 import { cn } from '@/lib/utils';
 import { useState } from 'react';
-import LogCard from '../../features/homepage/LogCard';
+import LogCard from '../../features/home/LogCard';
 import { Skeleton } from '../ui/skeleton';
 
 const MainPageLogCardList = () => {
