+++ src/features/Home/TypingText.tsx
