+++ src/features/register-page/CoverImageInput.tsx
