+++ src/features/register-page/ImagePreviewItem.tsx
