+++ src/features/register-page/LogWriteBar.tsx
