+++ src/features/register-page/RegisterBar.tsx
