+++ src/features/register-page/RegisterSearchBar.tsx
