+++ src/features/register-page/SelectedPlacePreview.tsx
