+++ src/pages/detail-page/PlacesCollectionPage.tsx
@@ -1,4 +1,4 @@
-import LogCard from '@/features/homepage/LogCard';
+import LogCard from '@/features/home/LogCard';
 import useLog from '@/hooks/queries/log/useLog';
 import { PlaceInLog } from '@/services/apis/types/logAPI.type';
 import { X } from 'lucide-react';
