+++ src/pages/detail-page/components/PlaceItem.tsx
