+++ src/pages/home/index.tsx
@@ -3,8 +3,8 @@ import ArrowIcon from '@/components/Icons/ArrowIcon';
 import MainPageLogCardList from '@/components/LogCard/MainPageLogCardList';
 import { Button } from '@/components/ui/button';
 import { REGISTER_SELECT } from '@/constants/pathname';
-import MainHero from '@/features/homepage/MainHero';
-import TypingText from '@/features/homepage/TypingText';
+import MainHero from '@/features/home/MainHero';
+import TypingText from '@/features/home/TypingText';
 import { useNavigate } from 'react-router-dom';
 
 const HomePage = () => {
