+++ src/pages/register-page/EditPage.tsx
@@ -4,9 +4,9 @@ import { Button } from '@/components/ui/button';
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Textarea } from '@/components/ui/textarea';
-import LogCoverEditInput from '@/features/editPage/LogCoverEditInput';
-import LogEditBar from '@/features/editPage/LogEditBar';
-import PlaceEditFormItem from '@/features/editPage/PlacEditFormItem';
+import LogCoverEditInput from '@/features/edit-page/LogCoverEditInput';
+import LogEditBar from '@/features/edit-page/LogEditBar';
+import PlaceEditFormItem from '@/features/edit-page/PlacEditFormItem';
 import useLog from '@/hooks/queries/log/useLog';
 import { cn } from '@/lib/utils';
 import api from '@/services/apis/api';
