+++ src/pages/register-page/LogWritePage.tsx
@@ -3,9 +3,9 @@ import ModifyDrawer from '@/components/Drawer/ModifyDrawer';
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Textarea } from '@/components/ui/textarea';
-import LogCoverImgInput from '@/features/registerpage/LogCoverImgInput';
-import LogWriteBar from '@/features/registerpage/LogWriteBar';
-import PlaceFormItem from '@/features/registerpage/PlaceFormItem';
+import LogCoverImgInput from '@/features/register-page/LogCoverImgInput';
+import LogWriteBar from '@/features/register-page/LogWriteBar';
+import PlaceFormItem from '@/features/register-page/PlaceFormItem';
 import { cn } from '@/lib/utils';
 import api from '@/services/apis/api';
 import { Log, PresignUrlResponse } from '@/services/apis/types/registerAPI.type';
