+++ src/pages/register-page/MapPage.tsx
@@ -2,9 +2,9 @@ import Loading from '@/components/Loading';
 import { Button } from '@/components/ui/button';
 import { REGISTER_DETAILS, REGISTER_SEARCH } from '@/constants/pathname';
 import { useKakaoMap } from '@/contexts/KakaoMap.context';
-import RegisterSearchBar from '@/features/registerpage/RegisterSearchBar';
-import SearchResultDrawer from '@/features/registerpage/SearchResultDrawer';
-import SelectedPlacePreview from '@/features/registerpage/SelectedPlacePreview';
+import RegisterSearchBar from '@/features/register-page/RegisterSearchBar';
+import SearchResultDrawer from '@/features/register-page/SearchResultDrawer';
+import SelectedPlacePreview from '@/features/register-page/SelectedPlacePreview';
 import { cn } from '@/lib/utils';
 import { useRegisterStore } from '@/store/registerStore';
 import { FormEvent, useEffect, useRef, useState } from 'react';
