+++ src/pages/register-page/NewPlacePage.tsx
